fix: use CJS forge config
